chore: generate and maintain in-repo module documentation
The generate-terraform-docs workflow will use the terraform-docs/gh-actions action to generate Terraform documentation based on the module's provider requirements, variables, and outputs.

The peter-evans/create-pull-request action will open new pull requests to ensure that the README.md of each module, including the root module, includes this generated content and that the content stays up to date as changes are made to the project.

Signed-off-by: Marques Johansson <mjohansson@equinix.com>